==> 06_mpra_anno.Rmd <==
---
title: "06: Annotation of MPRA results"
author: "felix.struebing@med.uni-muenchen.de"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE)
```

```{r}
library(tidyverse)
library(GenomicRanges)
library(rtracklayer)
library(broom)
source('ggplot_theme_FLS.R')
```

```{r}
pubfolder <- '/earth/public_data/'
datafolder <- '/earth/SNCA_MPRA/'
```

```{r}
res <- readRDS('04_export/mpra_results.rds')
res.gr <- makeGRangesFromDataFrame(res, keep.extra.columns = TRUE)
```

```{r}
table(res$significant)
```

```{r}
table(res$significant, res$logFC> 0.5)
```
```{r}
table(res$significant, res$logFC < -2.5)
table(res$significant, res$logFC > 2.5)
```

### Annotation with ENCODE

Load the ENCODE cCRE classification for hg38:

```{r}
ccres <- read_delim('/earth/public_data/encode3/encode_ccres_alltissues_hg38_220207.tsv', delim = '\t')
ccres <- makeGRangesFromDataFrame(ccres, seqnames.field = '#chrom', start.field = 'chromStart', end.field = 'chromEnd', keep.extra.columns = TRUE)
```

First, classify our MPRA hits according to the ENCODE conventions:

```{r}
hits <- findOverlaps(res.gr, ccres, ignore.strand = TRUE)
widths <- width(pintersect(res.gr[queryHits(hits)], ccres[subjectHits(hits)]))
ovs <- res.gr[queryHits(hits)]
ovs$ccre <- ccres[subjectHits(hits)]$ccre
ovs$ccre_name <- ccres[subjectHits(hits)]$name
ovs$overlap_width <- widths
length(res.gr)
length(ovs)
```

Some ROIs overlap more than one cCRE, let's filter for the largest overlap to reduce redundancy:

```{r}
table(duplicated(ovs))
ovs_dup <- ovs %>% 
  as.data.frame(.) %>% 
  mutate(dupe = ifelse(duplicated(seq), 'yes', 'no' )) %>%
  mutate(dupe = ifelse(duplicated(seq, fromLast = TRUE), 'yes', dupe )) %>%  # `duplicated` does not count the original occurrence
  mutate(locus = paste(seqnames, start, end, strand, sep = '_'))
duprem <- ovs_dup %>% filter(dupe == 'yes') %>%
  group_by(locus) %>%
  mutate(top_candidate = max(overlap_width)) %>% 
  filter(overlap_width == top_candidate)
```

```{r}
res_anno_nodupe <- ovs_dup %>% filter(dupe == 'no') %>% 
  bind_rows(., duprem)
table(duplicated(res_anno_nodupe$seq))
res_anno_nodupe <- res_anno_nodupe %>% 
  filter(!duplicated(seq)) %>% 
  dplyr::select(-dupe, -top_candidate) %>% 
  mutate(ccre_simple = str_remove(ccre, ',CTCF-bound'))
table(duplicated(res_anno_nodupe$seq))
```

```{r}
res_anno_nodupe <- res_anno_nodupe %>% 
  mutate(ccre_simple = case_when(
    grepl(pattern = 'dELS|pELS', x = ccre) ~ 'enhancer',
    grepl(pattern = 'PLS', x = ccre) ~ 'promoter',
    TRUE ~ ccre)) %>% 
    mutate(ctcf_bound = ifelse(grepl(pattern = 'CTCF', x = ccre), 'yes', 'no')
  )
```

```{r}
res_anno_nodupe %>% 
  dplyr::filter(significant == 'yes') %>% 
  dplyr::count(ccre_simple, ctcf_bound) %>% 
  dplyr::filter(ccre_simple %in% c('promoter', 'enhancer'))

```

```{r}
res <- res_anno_nodupe
res.gr <- makeGRangesFromDataFrame(res, keep.extra.columns = TRUE)
```

```{r}
table(res$significant)
```

```{r}
table(res$significant, res$logFC > 0.5)
```

```{r}
table(res$significant, res$logFC > 2.5)
table(res$significant, res$logFC < -2.5)
```

### Plots - batch effects

Define whether or not an RE overlaps repeats:

```{r}
repeats.gr <- read_delim(paste0(pubfolder, 'hg38_repeatMasker_UCSC'), delim = '\t') %>% 
  dplyr::select('genoName', 'genoStart', 'genoEnd', 'strand', 'repName', 'repFamily') %>% 
  makeGRangesFromDataFrame(., seqnames.field = 'genoName', start.field = 'genoStart', end.field = 'genoEnd', keep.extra.columns = TRUE)
hits <- findOverlaps(res.gr, repeats.gr, select = 'arbitrary', minoverlap = 20L)
res$has_repeat <- hits
res$has_repeat <- ifelse(is.na(res$has_repeat), 'no', 'yes')
table(res$has_repeat)
```

Is there a relationship between strength of upregulation and repeat status or similar?

```{r}
res$is_strong <- ifelse(res$logFC > 2.5, 'strong', 'not_strong')
summary(lm(data = res, coverage ~ is_strong))
```

```{r}
res %>% ggplot(aes(x = logFC, y = -log10(pval), color = scale(log2(coverage)))) +
  geom_point(alpha = 0.7) +
  scale_color_continuous(type = 'viridis') +
  theme_FLS() +
  labs(color = 'Coverage')
```


```{r}
res$width = res$end - res$start
summary(lm(data = res, width ~ is_strong))
```
```{r}
res %>% ggplot(aes(x = logFC, y = -log10(pval), color = width)) +
  geom_point(alpha = 0.7) +
  scale_color_continuous(type = 'viridis') +
  labs(color = 'RE width') +
  theme_FLS()
```


```{r}
summary(lm(data = res, logFC ~ has_repeat))
```
```{r}
res %>% ggplot(aes(x = logFC, y = -log10(pval), color = has_repeat)) +
  geom_point(alpha = 0.7) +
  theme_FLS()
```

<!-- ```{r} -->
<!-- summary(lm(data = res, logFC ~ seqnames)) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- res %>% ggplot(aes(x = logFC, y = -log10(pval), color = seqnames)) + -->
<!--   geom_point(alpha = 0.7) + -->
<!--   theme_FLS() -->
<!-- ``` -->

Remove spurious cCRE annotations:

```{r}
res_plot <- res %>% dplyr::filter(!grepl(x = ccre_simple, pattern = 'DNase')) %>% 
  mutate(ccre_simple = ifelse(ccre_simple == 'CTCF-only,CTCF-bound', 'CTCF-only', ccre_simple)) %>% 
  mutate(mycol = as.factor(case_when(
    ccre_simple == 'enhancer' ~ '#FFCB4B',
    ccre_simple == 'promoter' ~ '#FF1511',
    ccre_simple == 'CTCF-only' ~ '#01B2E9'
  )))
res_plot$ccre_simple = factor(x = res_plot$ccre_simple, levels = c('CTCF-only', 'promoter', 'enhancer'))
```


```{r fig.width = 8, fig.height = 8}
res_plot %>% 
  ggplot() +
  geom_point(data = . %>% dplyr::filter(significant == 'yes'), aes(x = logFC, y = -log10(pval), color = ccre_simple, shape = ctcf_bound), alpha = 0.7, size = 3) +
  geom_point(data = . %>% dplyr::filter(significant == 'no'), color = 'grey50', alpha = 0.3, aes(x = logFC, y = -log10(pval))) +
  scale_color_manual(values = c('#01B2E9', '#FF1511', '#FFA631'), labels = c('CTCF-only', 'Promoter', 'Enhancer')) +
  guides(color = guide_legend(title = ''), shape = guide_legend(title = 'CTCF-bound')) +
  labs(title = 'Differentially active REs: SNCA vs. GFP') +
  theme_FLS() +
  theme(legend.position = 'top')
```

```{r}
table(res$significant, res$ccre_simple)
```
```{r}
res_plot %>% group_by(significant, ctcf_bound, ccre_simple) %>% 
  tally()
```

Is log2-FC related to ctcf occupancy?

```{r}
res %>% dplyr::filter(significant == 'yes') %>% lm(logFC ~ ctcf_bound, data  = .) %>% glance()
```

Are promoters less activated than enhancer?

```{r}
res %>% dplyr::filter(significant == 'yes' & ccre_simple %in% c('enhancer', 'promoter')) %>% lm(logFC ~ ccre_simple, data = .) %>% glance()
```

<!-- ```{r} -->
<!-- res %>% dplyr::filter(significant == 'yes' & ccre_simple %in% c('enhancer', 'promoter')) %>%  -->
<!--   mutate(ccre_simple = ifelse(ccre_simple == 'enhancer', 1, 0)) %>%  -->
<!--   ggplot(aes(x = logFC, y = ccre_simple)) + -->
<!--   geom_point() +  -->
<!--   stat_smooth(method="glm", color="green", se=FALSE, -->
<!--                 method.args = list(family='binomial')) -->
<!-- ``` -->



Promoter annotation:

```{r}
res %>% dplyr::filter(ccre_simple == 'promoter') %>% arrange(logFC, fdr)
```



Export results to bed format:

```{r}
mypal <- colorRampPalette(c('yellow', 'white', 'blue'))
# res.gr$itemRgb <- '#7F7F7F'
res.gr$itemRgb <- mypal(20)[cut(res.gr$logFC, breaks = 20)]
res.gr$score <- res.gr$fdr
res.gr_exp <- res.gr[res.gr$coverage > 100]
export.bed(res.gr_exp, '06_export/results_colored_by_logfc.bed')
```

```{r}
res_simple <- res %>% 
  mutate(locus = paste0(seqnames, ':', start, '-', end)) %>% 
  dplyr::select(locus, strand, seq, fdr, logFC, ccre)
res_simple %>% 
  dplyr::filter(seq == 'TCAGTCGCAGCCCCA')
```

```{r eval = FALSE}
saveRDS(res, '06_export/results_filtered.rds')
```
